style: reorganise pages
