+++ lib/main.dart
@@ -12,15 +12,15 @@ import 'notifiers/line_line_statuses_filter_change_notifier.dart';
 import 'notifiers/line_predictions_filter_change_notifier.dart';
 import 'notifiers/lines_filter_change_notifier.dart';
 import 'notifiers/tfl_api_change_notifier.dart';
-import 'pages/lines/line_disruption_page.dart';
+import 'pages/line_disruptions/line_disruption_page.dart';
+import 'pages/line_routes/line_route_page.dart';
+import 'pages/line_statuses/line_status_page.dart';
 import 'pages/lines/line_line_disruptions_page.dart';
 import 'pages/lines/line_line_routes_page.dart';
 import 'pages/lines/line_line_statuses_page.dart';
 import 'pages/lines/line_page.dart';
 import 'pages/lines/line_predictions_page.dart';
-import 'pages/lines/line_route_page.dart';
 import 'pages/lines/line_route_sequences_page.dart';
-import 'pages/lines/line_status_page.dart';
 import 'pages/lines/line_stop_points_page.dart';
 import 'pages/lines/lines_page.dart';
 import 'pages/predictions/prediction_page.dart';
