+++ lib/material/list_tile.dart
@@ -2,10 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:tfl_api_client/tfl_api_client.dart' as tflApiClient;
 
-import '../pages/lines/line_disruption_page.dart';
+import '../pages/line_disruptions/line_disruption_page.dart';
+import '../pages/line_routes/line_route_page.dart';
+import '../pages/line_statuses/line_status_page.dart';
 import '../pages/lines/line_page.dart';
-import '../pages/lines/line_route_page.dart';
-import '../pages/lines/line_status_page.dart';
 import '../pages/predictions/prediction_page.dart';
 import '../pages/route_sequences/route_sequence_page.dart';
 import '../pages/stop_points/stop_point_page.dart';
