+++ lib/pages/line_statuses/line_status_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:tfl_api_client/tfl_api_client.dart';
 
-import '../widgets/text.dart';
+import '../../widgets/text.dart';
 
 class LineStatusPage extends StatefulWidget {
   static const route = '/line_statuses/:id';
